Prepare Order Handling

+++ ACMESharpCore/ACMESharpCore.psm1
@@ -24,9 +24,21 @@ $script:PSModuleRoot = $PSScriptRoot
 
 if (!(Test-Path -Path "$script:PSModuleRoot\FullModule.ps1")) {
     $classPath = "$script:PSModuleRoot\internal\classes";
+    $classes = @(
+        "AcmeHttpResponse",
+        "AcmeObject",
+        "AcmeDirectory",
+        "AcmeAccount",
+        "AcmeIdentifier",
+        "AcmeOrder"
+    )
+    
+    $classMergeFile = "$classPath\..\AllClasses.ps1";
+    if(Test-Path $classMergeFile) {
+        Clear-Content "$classPath\..\AllClasses.ps1"
+    }
 
-    Clear-Content "$classPath\..\AllClasses.ps1"
-    $classes | ForEach-Object { Get-Content "$classPath\$_.ps1" } | Set-Content "$classPath\..\AllClasses.ps1"
+    $classes | ForEach-Object { Get-Content "$classPath\$_.ps1" } | Set-Content $classMergeFile
 
     #TODO: This is a workaround for loading the functions. Needed it, because using a class inside the ctor of another class was not possible.
     . Import-ModuleFile "$classPath\..\AllClasses.ps1";
